[deckhouse] move the public-domain helper under hooks/lib
Follow the layout the repository already uses for a Go package a module's hooks
share, and keep it importable from the template tests.
