+++ modules/002-deckhouse/hooks/lib/publicdomain/publicdomain.go
